--- hello/state.go
@@ -1,39 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package hello
-
-import (
-	"sync"
-)
-
-// taskStore provides a mechanism to store and retrieve
-// task handles given a string identifier. The ID should
-// be unique per task
-type taskStore struct {
-	store map[string]*taskHandle
-	lock  sync.RWMutex
-}
-
-func newTaskStore() *taskStore {
-	return &taskStore{store: map[string]*taskHandle{}}
-}
-
-func (ts *taskStore) Set(id string, handle *taskHandle) {
-	ts.lock.Lock()
-	defer ts.lock.Unlock()
-	ts.store[id] = handle
-}
-
-func (ts *taskStore) Get(id string) (*taskHandle, bool) {
-	ts.lock.RLock()
-	defer ts.lock.RUnlock()
-	t, ok := ts.store[id]
-	return t, ok
-}
-
-func (ts *taskStore) Delete(id string) {
-	ts.lock.Lock()
-	defer ts.lock.Unlock()
-	delete(ts.store, id)
-}
