+++ main.go
@@ -5,7 +5,7 @@ package main
 
 import (
 	// TODO: update the path below to match your own repository
-	"github.com/andrewesweet/nomad-driver-milo/hello"
+	"github.com/andrewesweet/nomad-driver-milo/milo"
 
 	"github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/nomad/plugins"
@@ -18,5 +18,5 @@ func main() {
 
 // factory returns a new instance of a nomad driver plugin
 func factory(log hclog.Logger) interface{} {
-	return hello.NewPlugin(log)
+	return milo.NewPlugin(log)
 }
